backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
boto3
botocore
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==4.2.27
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
gunicorn==23.0.0
h11==0.16.0
idna==3.11
jmespath==1.0.1
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
requests
s3transfer
selenium<4.10.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3<2.0.0
webdriver-manager==4.0.2
websocket-client==1.9.0
whitenoise==6.11.0
wsproto==1.2.0